Cycle through old issues after processing all new ones
refactor auto discover system and rewrite so it orders items like: All issues/PRs that aren't in the database, or have been updated since they were last added to the database, by most recent to oldest. Then all other issues/PRs, including those that haven't updated and are already tracked. This is so we keep cycling through the entire issue set and use all the triages we are allowed to. Implement this in the simplest way possible, preferably in a single queue/list. Remember this only applies to the auto discover mode which kicks in when no targets are specified. Add basic tests as well.